refactor(Attach, Attachment): rename Attach to Attachment (#1500)
issue - #1270

Closes gazprom-neft/consta-uikit#1270
