--- src/components/Attach/Attach.css
@@ -1,58 +1,0 @@
-.Attach {
-  display: flex;
-  flex-direction: row;
-  box-sizing: border-box;
-  width: 100%;
-  padding: var(--space-2xs) var(--space-xs);
-  background: var(--color-control-bg-clear);
-  border-radius: var(--control-radius);
-  text-decoration: none;
-  transition: background 0.2s ease;
-
-  &-File {
-    flex: none;
-    margin: var(--space-2xs) var(--space-xs) var(--space-2xs) 0;
-    transition: opacity 0.2s;
-
-    &_error {
-      opacity: 0.4;
-    }
-  }
-
-  &-Button {
-    flex: none;
-    opacity: 0;
-    transition: opacity 0.2s ease;
-  }
-
-  &-Content {
-    overflow: hidden;
-    flex: 1;
-    margin-top: var(--space-3xs);
-    word-wrap: break-word;
-
-    &:not(:last-child) {
-      margin-right: var(--space-xs);
-    }
-  }
-
-  &_withAction {
-    cursor: pointer;
-
-    &:hover {
-      background: var(--color-control-bg-clear-hover);
-    }
-  }
-
-  &_withOnButtonAction {
-    padding: var(--space-2xs) var(--space-2xs) var(--space-2xs) var(--space-xs);
-
-    &:hover {
-      background: var(--color-control-bg-clear-hover);
-
-      .Attach-Button {
-        opacity: 1;
-      }
-    }
-  }
-}
