+++ src/components/Attach/Attach.tsx
@@ -1,103 +1,14 @@
-import './Attach.css';
+import { Attachment, cnAttachment } from '../Attachment/Attachment';
 
-import React from 'react';
+export { Attachment, cnAttachment } from '../Attachment/Attachment';
 
-import { IconProps } from '../../icons/Icon/Icon';
-import { cn } from '../../utils/bem';
-import { forwardRefWithAs } from '../../utils/types/PropsWithAsAttributes';
-import { Button } from '../Button/Button';
-import { File } from '../File/File';
-import { Text } from '../Text/Text';
+/**
+ * @deprecated since version 3.2.0 use Attachment
+ */
+const Attach = Attachment;
+/**
+ * @deprecated since version 3.2.0 cnAttachment
+ */
+const cnAttach = cnAttachment;
 
-type Props = {
-  fileExtension?: string;
-  loading?: boolean;
-  fileName?: string;
-  fileDescription?: string;
-  loadingProgress?: number;
-  errorText?: string;
-  loadingText?: string;
-  onButtonClick?: React.EventHandler<React.MouseEvent>;
-  buttonIcon?: React.FC<IconProps>;
-  buttonTitle?: string;
-  withAction?: boolean;
-  className?: string;
-  children?: never;
-};
-
-export const cnAttach = cn('Attach');
-
-export const Attach = forwardRefWithAs<Props>((props, ref) => {
-  const {
-    className,
-    as = 'div',
-    fileExtension,
-    loading,
-    fileName,
-    buttonIcon,
-    onButtonClick,
-    errorText,
-    loadingText = 'Loading',
-    loadingProgress,
-    fileDescription,
-    onClick,
-    withAction: withActionProp,
-    buttonTitle,
-    ...otherProps
-  } = props;
-  const Tag = as as string;
-  const withAction = withActionProp || Boolean(onClick);
-  const withOnButtonAction = Boolean(onButtonClick);
-
-  return (
-    <Tag
-      {...otherProps}
-      onClick={onClick}
-      className={cnAttach({ withAction, withOnButtonAction }, [className])}
-      ref={ref}
-    >
-      <File
-        className={cnAttach('File', { error: Boolean(errorText) })}
-        extension={fileExtension}
-        loading={loading}
-        loadingWithProgressSpin
-        loadingProgress={loadingProgress}
-        size="s"
-      />
-      <div className={cnAttach('Content')}>
-        {fileName && (
-          <Text className={cnAttach('FileName')} size="s" view="primary" lineHeight="xs">
-            {fileName}
-          </Text>
-        )}
-        {fileDescription && !loading && (
-          <Text className={cnAttach('FileDescription')} size="xs" lineHeight="xs" view="ghost">
-            {fileDescription}
-          </Text>
-        )}
-        {loadingText && loading && (
-          <Text className={cnAttach('LoadingText')} size="xs" lineHeight="xs" view="ghost">
-            {loadingProgress ? `${loadingText} ${loadingProgress}%` : `${loadingText}...`}
-          </Text>
-        )}
-        {errorText && (
-          <Text className={cnAttach('ErrorText')} size="xs" lineHeight="xs" view="alert">
-            {errorText}
-          </Text>
-        )}
-      </div>
-      {withOnButtonAction && (
-        <Button
-          className={cnAttach('Button')}
-          as="span"
-          onlyIcon
-          iconLeft={buttonIcon}
-          onClick={onButtonClick}
-          title={buttonTitle}
-          size="xs"
-          view="clear"
-        />
-      )}
-    </Tag>
-  );
-});
+export { Attach, cnAttach };
