--- src/components/Attach/__stories__/Attach.stories.css
@@ -1,3 +1,0 @@
-.AttachStories {
-  max-width: 320px;
-}
