--- src/components/Attach/__stories__/examples/AttachExampleAs/AttachExampleAs.css
@@ -1,3 +1,0 @@
-.AttachExampleAs {
-  max-width: 320px;
-}
