--- src/components/Attach/__stories__/examples/AttachExampleBasics/AttachExampleBasics.css
@@ -1,3 +1,0 @@
-.AttachExampleBasics {
-  max-width: 320px;
-}
