--- src/components/Attach/__stories__/examples/AttachExampleDescription/AttachExampleDescription.css
@@ -1,3 +1,0 @@
-.AttachStories--ExampleDescription {
-  max-width: 320px;
-}
