--- src/components/Attach/__stories__/examples/AttachExampleError/AttachExampleError.css
@@ -1,3 +1,0 @@
-.AttachExampleError {
-  max-width: 320px;
-}
