--- src/components/Attach/__stories__/examples/AttachExampleEvents/AttachExampleEvents.css
@@ -1,3 +1,0 @@
-.AttachExampleEvents {
-  max-width: 320px;
-}
