--- src/components/Attach/__stories__/examples/AttachExampleIcon/AttachExampleIcon.css
@@ -1,3 +1,0 @@
-.AttachStories--ExampleName {
-  max-width: 320px;
-}
