--- src/components/Attach/__stories__/examples/AttachExampleLoading/AttachExampleLoading.css
@@ -1,3 +1,0 @@
-.AttachStories--ExampleLoading {
-  max-width: 320px;
-}
