--- src/components/Attach/__stories__/examples/AttachExampleName/AttachExampleName.css
@@ -1,3 +1,0 @@
-.AttachStories--ExampleIcon {
-  max-width: 320px;
-}
