+++ src/components/Attachment/__stories__/Attachment.stories.css
@@ -1,0 +1,3 @@
+.AttachmentStories {
+  max-width: 320px;
+}
