+++ src/components/Attachment/__stories__/examples/AttachmentExampleAs/AttachmentExampleAs.css
@@ -1,0 +1,3 @@
+.AttachmentExampleAs {
+  max-width: 320px;
+}
