+++ src/components/Attachment/__stories__/examples/AttachmentExampleBasics/AttachmentExampleBasics.css
@@ -1,0 +1,3 @@
+.AttachmentExampleBasics {
+  max-width: 320px;
+}
