+++ src/components/Attachment/__stories__/examples/AttachmentExampleDescription/AttachmentExampleDescription.css
@@ -1,0 +1,3 @@
+.AttachmentStories--ExampleDescription {
+  max-width: 320px;
+}
