+++ src/components/Attachment/__stories__/examples/AttachmentExampleError/AttachmentExampleError.css
@@ -1,0 +1,3 @@
+.AttachmentExampleError {
+  max-width: 320px;
+}
