+++ src/components/Attachment/__stories__/examples/AttachmentExampleEvents/AttachmentExampleEvents.css
@@ -1,0 +1,3 @@
+.AttachmentExampleEvents {
+  max-width: 320px;
+}
