+++ src/components/Attachment/__stories__/examples/AttachmentExampleIcon/AttachmentExampleIcon.css
@@ -1,0 +1,3 @@
+.AttachmentStories--ExampleName {
+  max-width: 320px;
+}
