+++ src/components/Attachment/__stories__/examples/AttachmentExampleLoading/AttachmentExampleLoading.css
@@ -1,0 +1,3 @@
+.AttachmentStories--ExampleLoading {
+  max-width: 320px;
+}
