+++ src/components/Attachment/__stories__/examples/AttachmentExampleName/AttachmentExampleName.css
@@ -1,0 +1,3 @@
+.AttachmentStories--ExampleIcon {
+  max-width: 320px;
+}
