+++ src/components/File/__stories__/File.docs.mdx
@@ -8,7 +8,7 @@ import { FileExampleSize } from './examples/FileExampleSize/FileExampleSize';
 когда нужно показать файл — один или в списке.
 
 > Для файла, который загружает пользователь, есть специальный компонент —
-> [Attach](?path=/story/components-attach--playground)
+> [Attachment](?path=/story/components-attachment--playground)
 
 - [Иконка](#иконка)
 - [Размер иконки](#размер-иконки)
